Finished all endpoints and Http Responses, and Added Tests for jpa methods.

+++ src/main/java/com/angel/inventorycrud/controller/InventoryRestController.java
@@ -2,6 +2,7 @@ package com.angel.inventorycrud.controller;
 
 import com.angel.inventorycrud.entity.Item;
 import com.angel.inventorycrud.service.InventoryServiceImpl;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -25,25 +26,42 @@ public class InventoryRestController {
 
     @GetMapping("/items/{id}")
     public ResponseEntity<Item> displayItemById(@PathVariable int id){
-        ResponseEntity<Item> dbItem = inventoryService.find(id);
-        return dbItem;
+        Item dbItem =  inventoryService.find(id);
+        if (dbItem!=null){
+        return  new ResponseEntity<>(dbItem,HttpStatus.OK);
+        }
+
+        return ResponseEntity.notFound().build();
     }
 
 
     @GetMapping("/items")
     public ResponseEntity<List<Item>> displayItems(@RequestParam(name = "state",required = false) String state){
-         return inventoryService.findAll(state);
+         List<Item> dbItems = inventoryService.findAll(state);
+         if (dbItems.isEmpty()) return ResponseEntity.notFound().build();
+         return new ResponseEntity<>(dbItems,HttpStatus.OK);
 
     }
     @PostMapping("/items")
-    public Item addItem(@RequestBody Item theItem){
+    public ResponseEntity<Item> addItem(@RequestBody Item theItem){
+
         Item dbItem = inventoryService.save(theItem);
-        return dbItem;
+        if(dbItem==null){
+            return ResponseEntity.badRequest().build();
+
+        }
+        return new ResponseEntity<>(dbItem,HttpStatus.CREATED);
     }
 
     @DeleteMapping("/items/{id}")
-    public void removeItem(@PathVariable int id){
+    public ResponseEntity<Object> removeItem(@PathVariable int id){
+
+       Item dbItem = inventoryService.find(id);
+        if (dbItem==null){
+            return   ResponseEntity.notFound().build();
+        }
         inventoryService.remove(id);
+        return ResponseEntity.noContent().build();
     }
